Flatten Binary Tree(recursive solution)
This is faster than the constant space solution as we are visiting all the nodes only once and we are encoutering the left and right tail for each node, then we are connecting out left tail to root->right and we are returning the right tail which becomes the left tail of the next root. 
This process continues both on left subtree and right subtree.